backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
distro==1.9.0
fastapi==0.134.0
fonttools==4.61.1
frozenlist==1.8.0
future==1.0.0
geocoder==1.38.1
geographiclib==2.1
geopy==2.4.1
groq==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
kiwisolver==1.4.9
matplotlib==3.10.8
multidict==6.7.1
numpy==2.4.2
opencv-python==4.13.0.92
packaging==26.0
pillow==12.1.1
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5

PyJWT==2.11.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
ratelim==0.1.6
reportlab==4.4.10
requests==2.32.5
six==1.17.0
sniffio==1.3.1
starlette==0.52.1
twilio==9.10.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
yarl==1.22.0